requirements.txt
distlib==0.3.7
filelock==3.12.4
numpy==1.26.2
numpy-stl==3.1.0
platformdirs==3.11.0
PyOpenGL==3.1.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
python-utils==3.8.1
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5